+++ packages/pi-software-factory/src/extension.ts
@@ -1,4 +1,4 @@
-import type { ExtensionAPI } from "@oh-my-pi/pi-coding-agent";
+import type { ExtensionAPI } from "@oh-my-pi/pi-coding-agent/extensibility/extensions/types";
 import { runFactoryDoctor } from "./doctor";
 import { scaffoldFactory } from "./scaffold";